fix: configured project to meet project goals
